Changed sig_amp_cuts to be a Float

+++ ctapipe/image/charge_extractors.py
@@ -4,7 +4,7 @@ Charge extraction algorithms to reduce the image to one value per pixel
 
 from abc import abstractmethod
 import numpy as np
-from traitlets import Int, CaselessStrEnum
+from traitlets import Int, CaselessStrEnum, Float
 from ctapipe.core import Component, Factory
 
 __all__ = ['ChargeExtractorFactory', 'FullIntegrator', 'SimpleIntegrator',
@@ -454,14 +454,14 @@ class SimpleIntegrator(WindowIntegrator):
 
 class PeakFindingIntegrator(WindowIntegrator):
     name = 'PeakFindingIntegrator'
-    sig_amp_cut_HG = Int(None, allow_none=True,
-                         help='Define the cut above which a sample is '
-                              'considered as significant for PeakFinding '
-                              'in the HG channel').tag(config=True)
-    sig_amp_cut_LG = Int(None, allow_none=True,
-                         help='Define the cut above which a sample is '
-                              'considered as significant for PeakFinding '
-                              'in the LG channel').tag(config=True)
+    sig_amp_cut_HG = Float(None, allow_none=True,
+                           help='Define the cut above which a sample is '
+                                'considered as significant for PeakFinding '
+                                'in the HG channel').tag(config=True)
+    sig_amp_cut_LG = Float(None, allow_none=True,
+                           help='Define the cut above which a sample is '
+                                'considered as significant for PeakFinding '
+                                'in the LG channel').tag(config=True)
 
     def __init__(self, config, tool, **kwargs):
         """
@@ -728,16 +728,16 @@ class ChargeExtractorFactory(Factory):
                                'PeakFindingIntegrators.').tag(config=True)
     t0 = Int(0, help='Define the peak position for all pixels. '
                      'Only applicable to SimpleIntegrators.').tag(config=True)
-    sig_amp_cut_HG = Int(None, allow_none=True,
-                         help='Define the cut above which a sample is '
-                              'considered as significant for PeakFinding '
-                              'in the HG channel. Only applicable to '
-                              'PeakFindingIntegrators.').tag(config=True)
-    sig_amp_cut_LG = Int(None, allow_none=True,
-                         help='Define the cut above which a sample is '
-                              'considered as significant for PeakFinding '
-                              'in the LG channel. Only applicable to '
-                              'PeakFindingIntegrators.').tag(config=True)
+    sig_amp_cut_HG = Float(None, allow_none=True,
+                           help='Define the cut above which a sample is '
+                                'considered as significant for PeakFinding '
+                                'in the HG channel. Only applicable to '
+                                'PeakFindingIntegrators.').tag(config=True)
+    sig_amp_cut_LG = Float(None, allow_none=True,
+                           help='Define the cut above which a sample is '
+                                'considered as significant for PeakFinding '
+                                'in the LG channel. Only applicable to '
+                                'PeakFindingIntegrators.').tag(config=True)
     lwt = Int(0, help='Weight of the local pixel (0: peak from neighbours '
                       'only, 1: local pixel counts as much as any neighbour). '
                       'Only applicable to '
